Version Finale Backend

+++ play-service/src/main/java/ch/unige/PlayResource.java
@@ -1,16 +1,12 @@
 package ch.unige;
 
-//import java.net.URI;
 import java.util.ArrayList;
 import java.util.List;
 
-//import javax.management.relation.RelationSupport;
 import javax.transaction.Transactional;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
-//import javax.ws.rs.DELETE;
 import javax.ws.rs.GET;
-//import javax.ws.rs.HeaderParam;
 import javax.ws.rs.POST;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
@@ -26,7 +22,6 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 
 import ch.unige.dao.LobbyDB;
 import ch.unige.dao.UserInLobbyDB;
-//import io.vertx.core.json.JsonObject;
 
 @Path("/play")
 public class PlayResource {
@@ -35,8 +30,8 @@ public class PlayResource {
     @Path("users")
     @Produces(MediaType.APPLICATION_JSON)
     public Response getAllUsers() {
-        List<UserInLobbyDB> UIL = UserInLobbyDB.listAll();
-        return Response.ok(UIL).build();
+        List<UserInLobbyDB> uil = UserInLobbyDB.listAll();
+        return Response.ok(uil).build();
     }
 
     @GET
@@ -53,7 +48,7 @@ public class PlayResource {
 
         String userid = headers.getHeaderString("X-User");
 
-        UserInLobbyDB userExistant = UserInLobbyDB.getUser(userid);
+        var userExistant = UserInLobbyDB.getUser(userid);
         if (userExistant == null) {
             return Response.status(Response.Status.NOT_FOUND).build();
         } else {
@@ -72,45 +67,47 @@ public class PlayResource {
         try {
             initJSON = SecurityUtility.decrypt(initString);
         } catch (Exception e) {
-            String message = "Encryption non valide";
+            var message = "Encryption_non_valide";
             return Response.status(Response.Status.BAD_REQUEST).entity(message).build();
         }
 
-        StringToJSON jsonObj = null;
         try {
-            jsonObj = new ObjectMapper().readValue(initJSON, StringToJSON.class);
+            StringToJSON jsonObj = new ObjectMapper().readValue(initJSON, StringToJSON.class);
+
+            // Vérifier que le lobby n'est pas déjà en DB
+            LobbyDB lobbyExistant = LobbyDB.getLobby(jsonObj.token);
+            if (lobbyExistant != null) {
+                var message = "Lobby_déjà_enregistré_!";
+                return Response.status(Response.Status.UNAUTHORIZED).entity(message).build();
+            }
+
+            // Create players
+            for (String userID : jsonObj.listPlayer) {
+                UserInLobbyDB player = new UserInLobbyDB();
+                player.token = jsonObj.token;
+                player.userID = userID;
+                // add player to lobby
+                UserInLobbyDB.persist(player);
+            }
+
+            // Create and init lobby
+            LobbyDB lobby = new LobbyDB();
+            lobby.token = jsonObj.token;
+            for (int i = 0; i < 20; i++) {
+                lobby.sumScores.add(0);
+                lobby.numberVotes.add(0);
+            }
+
+            // Add lobby dans la DB
+            LobbyDB.persist(lobby);
+            return Response.created(null).build();
         } catch (JsonMappingException e) {
             e.printStackTrace();
         } catch (JsonProcessingException e) {
             e.printStackTrace();
         }
-
-        // Vérifier que le lobby n'est pas déjà en DB
-        LobbyDB lobbyExistant = LobbyDB.getLobby(jsonObj.token);
-        if (lobbyExistant != null) { 
-            String message = "Lobby déjà enregistré !";
-            return Response.status(Response.Status.UNAUTHORIZED).entity(message).build();
-        }
-
-        // Create players
-        for (String userID : jsonObj.listPlayer) {
-            UserInLobbyDB player = new UserInLobbyDB();
-            player.token = jsonObj.token;
-            player.userID = userID;
-            UserInLobbyDB.persist(player);
-        }
-
-        // Create and init lobby
-        LobbyDB lobby = new LobbyDB();
-        lobby.token = jsonObj.token;
-        for (int i = 0; i < 20; i++) {
-            lobby.sumScores.add(0);
-            lobby.numberVotes.add(0);
-        }
-
-        // Add lobby dans la DB
-        LobbyDB.persist(lobby);
-        return Response.created(null).build();
+        var message = "Erreur_avec_le_message_JSON_!";
+        return Response.status(Response.Status.BAD_REQUEST).entity(message).build();
     }
 
     @POST
@@ -120,41 +117,36 @@ public class PlayResource {
     public Response GameScore(@PathParam("MOVIEID") int movie_id, @PathParam("SCORE") int score,
             @Context HttpHeaders headers) {
 
-        String userid = headers.getHeaderString("X-User");
-
+        var userid = headers.getHeaderString("X-User");
         // movie_id valide ?
         if (movie_id < 0 || movie_id > 19 || movie_id != (int) movie_id) {
-            String message = "Le id du film n'est pas valide !";
+            var message = "Le_id_du_film_n'est_pas_valide !";
             return Response.status(Response.Status.BAD_REQUEST).entity(message).build();
         }
-
         // score valide ?
         if (score < 0 || score > 100 || score != (int) score) {
-            String message = "Le score attribué n'est pas valide !";
+            var message = "Le_score_attribué_n'est_pas_valide !";
             return Response.status(Response.Status.BAD_REQUEST).entity(message).build();
         }
-
         // Regarder si le sender est dans une partie
-        UserInLobbyDB UIL = UserInLobbyDB.getUser(userid);
-        if (UIL == null) {
-            String message = "Le joueur n'est pas dans une partie !";
+        var uil = UserInLobbyDB.getUser(userid);
+        if (uil == null) {
+            var message = "Le_joueur_n'est_pas_dans_une_partie_!";
             return Response.status(Response.Status.UNAUTHORIZED).entity(message).build();
         }
-
         // Get les attributs du sender
-        String token = UIL.getToken();
-        // boolean user_finished = UIL.getUser_finished();
-        ArrayList<Integer> votesID = UIL.getVotesID();
+        String token = uil.token;
+        ArrayList<Integer> votesID = uil.votesID;
 
         // Le joueur a déjà joué toutes ses cartes ?
         if (votesID.size() >= 20) {
-            String message = "Le joueur a déjà épuisé son nombre de votes !";
+            var message = "Le_joueur_a_déjà_épuisé_son_nombre_de_votes_!";
             return Response.status(Response.Status.UNAUTHORIZED).entity(message).build();
         }
 
         // Le joueur a déjà voté pour ce même film ?
         if (votesID.contains(movie_id)) {
-            String message = "Le joueur a déjà voté pour ce film !";
+            var message = "Le_joueur_a_déjà_voté_pour_ce_film_!";
             return Response.status(Response.Status.UNAUTHORIZED).entity(message).build();
         }
 
@@ -163,7 +155,7 @@ public class PlayResource {
 
         // Add le score côté user
         votesID.add(movie_id);
-        UIL.setVotesID(votesID);
+        uil.votesID = votesID;
 
         // Add le score côté lobby
         int currentScore = L.sumScores.get(movie_id);
@@ -180,16 +172,22 @@ public class PlayResource {
     @Produces(MediaType.TEXT_PLAIN)
     public Response UserQuitLobby(@Context HttpHeaders headers) {
 
-    	String userID = headers.getHeaderString("X-User");
+        var userID = headers.getHeaderString("X-User");
 
         // Regarder si le joueur est dans une partie
-        UserInLobbyDB UIL = UserInLobbyDB.getUser(userID);
-        if (UIL == null) {
-            String message = "Le joueur n'est pas dans une partie !";
+        var uil = UserInLobbyDB.getUser(userID);
+        if (uil == null) {
+            var message = "Le_joueur_n'est_pas_dans_une_partie_!";
             return Response.status(Response.Status.NO_CONTENT).entity(message).build();
         }
         UserInLobbyDB.deleteUser(userID);
-        String message = "Joueur retiré de la partie";
+        var message = "Joueur_retiré_de_la_partie";
+        // vérifier si il y a toujours qqn dans le lobby sinon delete lobby
+        String token = uil.token;
+        var user = UserInLobbyDB.getUserFromToken(token);
+        if (user == null) {
+            LobbyDB.deleteLobby(token);
+        }
         return Response.status(Response.Status.OK).entity(message).build();
     }
 
@@ -199,29 +197,63 @@ public class PlayResource {
     @Produces(MediaType.APPLICATION_JSON)
     public Response getResult(@Context HttpHeaders headers) {
 
-        String userid = headers.getHeaderString("X-User");
+        var userid = headers.getHeaderString("X-User");
 
-        // Vérifier si le user est dand un lobby et get le token
+        // Vérifier si le user est dans un lobby et get le token
         String token;
         try {
-            UserInLobbyDB UIL = UserInLobbyDB.getUser(userid);
-            token = UIL.getToken();
+            var uil = UserInLobbyDB.getUser(userid);
+            token = uil.token;
         } catch (Exception e) {
-            String message = "La requête n'est pas valide !";
+            var message = "La_requête_n'est_pas_valide_!";
             return Response.status(Response.Status.BAD_REQUEST).entity(message).build();
         }
 
-        // Vérifier que tous les joueurs de ce lobby ont fini de voter pour tous les
-        // films
         if (UserInLobbyDB.getStatus(token)) {
-            // Retourner le résultat
-            int movie_winner_id = LobbyDB.getMovieWinner(token);
-            UserInLobbyDB.deleteUsers(token);
-            LobbyDB.deleteLobby(token);
-            return Response.status(Response.Status.OK).entity(movie_winner_id).build();
+            var uil = UserInLobbyDB.getUser(userid);
+            int movieWinnerID;
+            if (uil.result == -1) {
+                // get result
+                movieWinnerID = LobbyDB.getMovieWinner(token);
+                // write results to all users
+                UserInLobbyDB.writeResultToDB(movieWinnerID, token);
+            } else {
+                movieWinnerID = uil.result;
+            }
+            // delete le user
+            UserInLobbyDB.deleteUser(userid);
+            // vérifier si il y a toujours qqn dans le lobby sinon delete lobby
+            var user = UserInLobbyDB.getUserFromToken(token);
+            if (user == null) {
+                LobbyDB.deleteLobby(token);
+            }
+            // return result
+            String message = "{\"id\":" + movieWinnerID + "}";
+            return Response.status(Response.Status.OK).entity(message).type(MediaType.APPLICATION_JSON).build();
         } else {
-            String message = "La partie n'est pas finie";
+            var message = "La_partie_n'est_pas_terminée";
             return Response.status(Response.Status.NO_CONTENT).entity(message).build();
         }
+    }
+
+    @GET
+    @Produces(MediaType.APPLICATION_JSON)
+    @Path("/checkAllFinish")
+
+    public Response gameDone(@Context HttpHeaders headers) {
+        var userid = headers.getHeaderString("X-User");
+
+        // Vérifier si le user est dans un lobby et get le token
+        String token;
+        try {
+            var uil = UserInLobbyDB.getUser(userid);
+            token = uil.token;
+        } catch (Exception e) {
+            var message = "La_requête_n'est_pas_valide_!";
+            return Response.status(Response.Status.BAD_REQUEST).entity(message).build();
+        }
+        boolean status = UserInLobbyDB.getStatus(token);
+        String message = "{\"fin\":" + status + "}";
+        return Response.status(Response.Status.OK).entity(message).type(MediaType.APPLICATION_JSON).build();
     }
 }
